requirements.txt
fastapi
uvicorn
pillow
rasterio
numpy

